refactor(lib): relocate sanitizeToolName to @typebot.io/lib for reuse
The builder needs the same MCP name normalization to detect tool-name
collisions at creation time. Move the pure helper into @typebot.io/lib
(a dependency of both the builder and mcp-tools) and re-export it from
mcp-tools so existing importers are unaffected.
